requirements.txt
asyncpg
aiogram==2.25.2
aiohttp
python-dotenv
pillow
Django
psycopg2-binary